Move component ConsultationAboutFeverNumberCard from 'components/cards' to 'components/index/CardsMonitoring'
