+++ components/index/CardsMonitoring/ConsultationAboutFeverNumberCard.vue
